payroll-automation/requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-aiplatform
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
watchdog==3.0.0
